KWN-3249:added check for Node version and added README.md file

+++ build.sh
@@ -1,9 +1,34 @@
-#!/bin/bash +x
+#!/bin/bash
+
+REQUIRED_NODE_MAJOR=20
 
 echo "---"
 echo "--- STARTING BUILD"
 echo "---"
-echo "--- Cleaning NPM node modules"
+# -----------------------------
+# Check if Node is installed
+# -----------------------------
+if ! command -v node >/dev/null 2>&1; then
+    echo "ERROR: Node.js is not installed."
+    exit 1
+fi
+# -----------------------------
+# Check Node version
+# -----------------------------
+NODE_VERSION=$(node -v)
+NODE_MAJOR=$(echo "$NODE_VERSION" | sed 's/v//' | cut -d. -f1)
+
+echo "Detected Node version: $NODE_VERSION"
+
+if [ "$NODE_MAJOR" -lt "$REQUIRED_NODE_MAJOR" ]; then
+    echo "ERROR: Node $REQUIRED_NODE_MAJOR or higher is required."
+    echo "Please install Node $REQUIRED_NODE_MAJOR+ and try again."
+    exit 1
+fi
+
+echo "Node version check passed"
+echo "---"
+echo "--- Deleting NPM node modules"
 echo "---"
 rm -rf node_modules/
 rm -rf baseline-analysis-task/node_modules/
@@ -37,7 +62,7 @@ npx tsc -p tsconfig-ext.json
 echo "---"
 echo "--- Packaging extension with TFX-CLI tool using vss-extension.json"
 echo "---"
-tfx extension create
+npx -y tfx-cli extension create
 
 echo "---"
 echo "--- BUILD FINISH"
